Move Cookie Generation in its own class.
